Started work on the first level. Adjusted cannons and made various fixes.
Started building the first proper level. Added a new launched enemy sprite that better fits the cannon, and added an offset field for how the projectiles are spawned. Switched breakable and disappearing blocks to use edge colliders to avoid ghost collisions. Made it so enemies die when going fully below the bottom of the level. Stopped the player from being set as defeated when already defeated.